backend_python/requirements.txt
fastapi[standard]
sqlmodel
aiosqlite
httpx
playwright
python-dotenv
